+++ src/main.c
@@ -18,9 +18,9 @@ int main(void){
 
   // initialize function manager and add some functions to test drawing
   struct ree_function_manager_t functions;
-  char* f = "f(x) = 1/x";
-  char* g = "g(x) = x^2 - 3x";
-  char* m = "m(x) = log(x)";
+  char* f = "f(x) = sqrt(x)";
+  char* g = "g(x) = -x^2 - 3x";
+  char* m = "m(x) = ln(x)";
   // factorials DON'T WORK but im too lazy to make a proper factorial function so next commit it is :)
 
   enum reh_error_code_e err = ree_initFunctionManager(&functions);